[Fix] #2035 Sonar 후속 — 카탈로그 도구 공통부 추출·복잡도 정리
SonarCloud quality gate(new_duplicated_lines_density 26.6% > 3%, applyChain
인지 복잡도 16 > 15, 중첩 템플릿 리터럴) 해소.

- tools/route-map/lib/station-catalog.mjs 신설: add-daegu-buksam-catalog(#2019)와
  add-station-catalog-entries(#1954)가 복제하던 공통부(결정적 station id 발급,
  planNewStation 행 planning, rideEdgePair, insertRow, previewChain, applyChain)를
  salt·region-agnostic 코어로 추출. createStationCatalog({salt, region})가 도구별
  salt/region에 바인딩한 함수를 반환하고, 두 도구는 이를 re-export만 한다.
- applyChain의 체인 항목 id 해소·이름·직결 엣지 존재 확인을 헬퍼(resolveEntryId,
  entryName, directEdgePresent)로 분리해 인지 복잡도를 기준 이하로 낮춤.
- buksam --check의 중첩 템플릿 리터럴을 existsLabel·occupiedNote 변수로 평탄화.

두 도구의 CLI·산출 계약 불변: add-daegu-buksam-catalog.test.mjs·
add-station-catalog-entries.test.mjs(총 22)·repository-contract(186)·route-map
전체 스위트 green. 동일 입력 팩에 대해 리팩터 전/후 산출 gz·index·--check
출력이 byte-identical임을 확인(데이터 불변, 코드 정리만).
