+++ tools/route-map/add-station-catalog-entries.mjs
@@ -25,8 +25,8 @@
 // KRIC 공식 레코드 반영 시 반드시 재사용한다(동일 물리역 id 분기 방지).
 //
 // 사용: node tools/route-map/add-station-catalog-entries.mjs [--pack …] [--index …] [--check]
-import { createHash } from "node:crypto";
 import { mutatePack, parsePackArgs } from "./station-surgery.mjs";
+import { createStationCatalog, insertRow, rideEdgePair } from "./lib/station-catalog.mjs";
 
 const REGION = "수도권";
 const INCHEON1 = "line-98718184f016"; // 수도권 인천1호선
@@ -39,13 +39,11 @@ const EVERLINE = "line-828f04afc588"; // 수도권 에버라인
 export const VERIFIED_PRESS_AT = Date.UTC(2026, 6, 11) / 1000;
 export const VERIFIED_KRIC_AT = Date.UTC(2026, 5, 30) / 1000;
 
-/** 신설 역의 결정적 station id — split 수술(#1789)의 해시 기반 id 정책과 동일. */
-export function newStationId(lineId, nameKo) {
-  const hex = createHash("sha256")
-    .update(`new-station:${lineId}:${nameKo}:1954`)
-    .digest("hex");
-  return `station-${hex.slice(0, 12)}`;
-}
+// 신설 역의 결정적 station id·행 planning·체인 반영은 공용 코어(#2035)를 쓴다.
+// salt는 split 수술(#1789)의 해시 기반 id 정책과 동일한 이슈 번호(1954).
+const catalog = createStationCatalog({ salt: 1954, region: REGION });
+export const { newStationId, planNewStation, previewChain, applyChain } = catalog;
+export { rideEdgePair };
 
 /** 신설 4역(공식 근거 첨부). 좌표 null = admitted 소스 미반영(날조 금지). */
 export const NEW_STATIONS = [
@@ -142,64 +140,6 @@ export const RENAME = {
   evidence: "KRIC 역사정보 2026-06-30 Y120 용인중앙시장(용인예술과학대)",
 };
 
-/**
- * 순수: 신설 역 스펙 → stations/station_lines 행. 기존 카탈로그 관례를 따른다
- * (normalized_name = name_ko 원문, station_code = line_sequence 문자열,
- * name_en/name_sub/platform_info 빈 문자열).
- */
-export function planNewStation(spec) {
-  const id = newStationId(spec.lineId, spec.name);
-  return {
-    station: {
-      id,
-      name_ko: spec.name,
-      name_en: "",
-      normalized_name: spec.name,
-      region: REGION,
-      latitude: spec.latitude,
-      longitude: spec.longitude,
-      data_quality_level: spec.dataQualityLevel,
-      data_source_type: spec.dataSourceType,
-      last_verified_at: spec.lastVerifiedAt,
-      name_sub: "",
-    },
-    stationLine: {
-      station_id: id,
-      line_id: spec.lineId,
-      station_code: String(spec.lineSequence),
-      line_sequence: spec.lineSequence,
-      platform_info: "",
-    },
-  };
-}
-
-/** 순수: 인접쌍 → 양방향 RIDE 엣지 2행(기존 RIDE 관례 필드값·id 포맷). */
-export function rideEdgePair(lineId, fromStationId, toStationId, lastVerifiedAt) {
-  const row = (a, b) => ({
-    id: `edge-${lineId}-${a}-${b}`,
-    from_node_id: `${a}:${lineId}`,
-    to_node_id: `${b}:${lineId}`,
-    duration_seconds: 120,
-    distance_meters: 0,
-    edge_type: "RIDE",
-    service_pattern: "LOCAL",
-    includes_stairs: 0,
-    stair_access_state: "UNKNOWN",
-    accessibility_status: "UNKNOWN",
-    reliability_score: 80,
-    facility_id: null,
-    last_verified_at: lastVerifiedAt,
-  });
-  return [row(fromStationId, toStationId), row(toStationId, fromStationId)];
-}
-
-function insertRow(db, table, row) {
-  const cols = Object.keys(row);
-  db.prepare(
-    `INSERT INTO ${table} (${cols.join(",")}) VALUES (${cols.map(() => "?").join(",")})`,
-  ).run(...cols.map((c) => row[c]));
-}
-
 function applyNewStation(db, spec) {
   const { station, stationLine } = planNewStation(spec);
   if (db.prepare("SELECT 1 FROM stations WHERE id=?").get(station.id)) {
@@ -228,72 +168,6 @@ function applyNewStation(db, spec) {
     lineSequence: spec.lineSequence,
     evidence: spec.evidence,
   };
-}
-
-/**
- * 읽기 전용: applyChain이 던지는 사전조건(체인 역 존재·removeDirect 직결 엣지 존재)을
- * --check 미리보기용으로 점검한다. applyChain과 동일한 id 해소 규칙을 쓴다.
- */
-export function previewChain(db, { lineId, chain, removeDirect }) {
-  const rows = [];
-  for (const entry of chain) {
-    const id = typeof entry === "string" ? newStationId(lineId, entry) : entry.id;
-    const name = typeof entry === "string" ? entry : entry.name;
-    const present = Boolean(db.prepare("SELECT 1 FROM stations WHERE id=?").get(id));
-    rows.push({ name, id, present });
-  }
-  for (const [a, b] of removeDirect) {
-    const present = Boolean(
-      db
-        .prepare(
-          "SELECT 1 FROM network_edges WHERE edge_type='RIDE' AND from_node_id=? AND to_node_id=?",
-        )
-        .get(`${a}:${lineId}`, `${b}:${lineId}`),
-    );
-    rows.push({ name: `직결 ${a}↔${b}`, id: `${a}:${lineId}`, present });
-  }
-  return rows;
-}
-
-export function applyChain(db, { lineId, chain, expectedMembers, removeDirect, lastVerifiedAt }) {
-  const ids = chain.map((entry) =>
-    typeof entry === "string" ? newStationId(lineId, entry) : entry.id,
-  );
-  for (const entry of chain) {
-    const id = typeof entry === "string" ? newStationId(lineId, entry) : entry.id;
-    if (!db.prepare("SELECT 1 FROM stations WHERE id=?").get(id)) {
-      const name = typeof entry === "string" ? entry : entry.name;
-      throw new Error(`${lineId}: 체인 역 없음 ${name}(${id})`);
-    }
-  }
-  const del = db.prepare(
-    "DELETE FROM network_edges WHERE edge_type='RIDE' AND from_node_id=? AND to_node_id=?",
-  );
-  let removed = 0;
-  for (const [a, b] of removeDirect) {
-    removed += del.run(`${a}:${lineId}`, `${b}:${lineId}`).changes;
-    removed += del.run(`${b}:${lineId}`, `${a}:${lineId}`).changes;
-  }
-  let inserted = 0;
-  for (let i = 0; i + 1 < ids.length; i += 1) {
-    for (const edge of rideEdgePair(lineId, ids[i], ids[i + 1], lastVerifiedAt)) {
-      const dup = db
-        .prepare(
-          "SELECT 1 FROM network_edges WHERE edge_type='RIDE' AND from_node_id=? AND to_node_id=?",
-        )
-        .get(edge.from_node_id, edge.to_node_id);
-      if (dup) continue;
-      insertRow(db, "network_edges", edge);
-      inserted += 1;
-    }
-  }
-  const members = db
-    .prepare("SELECT COUNT(*) c FROM station_lines WHERE line_id=?")
-    .get(lineId).c;
-  if (members !== expectedMembers) {
-    throw new Error(`${lineId}: 반영 후 멤버 수 ${members} ≠ 기대 ${expectedMembers}`);
-  }
-  return { lineId, removedEdges: removed, insertedEdges: inserted, members };
 }
 
 function applyRename(db, spec) {
